First draft of runlist UI
This version has persistence of JSON in CachedRunList model. It does not
follow chef JSON format so the editor does not bring big value, just
that I didn't have to code UI myself. We do the parsing so we keep
chef-like schema in DB.

Next step is to see if we could get the editor working with chef-like
schema. And add orchestration.